# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Configuration Change Tracker API
  version: 1.0.0
  description: API for logging, storing, and retrieving changes to domain-specific rules.

paths:
  /api/v1/changes:
    post:
      summary: Record a new configuration change
      operationId: createChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequest'
      responses:
        '201':
          description: Configuration change recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      summary: Retrieve change history with filters
      operationId: getChanges
      parameters:
        - name: type
          in: query
          description: Filter by the type of rule
          schema:
            $ref: '#/components/schemas/RuleType'
        - name: from
          in: query
          description: Start timestamp (ISO-8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End timestamp (ISO-8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of configuration changes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChangeResponse'

  /api/v1/changes/{id}:
    get:
      summary: Retrieve a specific change by ID
      operationId: getChangeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Change found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeResponse'
        '404':
          description: Change not found

components:
  schemas:
    RuleType:
      type: string
      enum: [CREDIT_LIMIT, APPROVAL_POLICY]

    ChangeOperation:
      type: string
      enum: [ADD, UPDATE, DELETE]

    ChangeRequest:
      type: object
      required: [type, operation, actor]
      properties:
        type:
          $ref: '#/components/schemas/RuleType'
        operation:
          $ref: '#/components/schemas/ChangeOperation'
        actor:
          type: string
          example: "admin-user-01"
        payload:
          $ref: '#/components/schemas/Rule'
          description: Payload for ADD or DELETE operations
        oldPayload:
          $ref: '#/components/schemas/Rule'
          description: Old payload for UPDATE operation
        newPayload:
          $ref: '#/components/schemas/Rule'
          description: New payload for UPDATE operation

    ChangeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/RuleType'
        operation:
          $ref: '#/components/schemas/ChangeOperation'
        actor:
          type: string
        timestamp:
          type: string
          format: date-time
        payload:
          $ref: '#/components/schemas/Rule'
          description: Payload for ADD or DELETE operations
        oldPayload:
          $ref: '#/components/schemas/Rule'
          description: Old payload for UPDATE operation
        newPayload:
          $ref: '#/components/schemas/Rule'
          description: New payload for UPDATE operation

    Rule:
      type: object
      description: Polymorphic rule based on the rule type
      discriminator:
        propertyName: type
        mapping:
          CREDIT_LIMIT: '#/components/schemas/CreditLimitRule'
          APPROVAL_POLICY: '#/components/schemas/ApprovalPolicyRule'
      oneOf:
        - $ref: '#/components/schemas/CreditLimitRule'
        - $ref: '#/components/schemas/ApprovalPolicyRule'

    CreditLimitRule:
      type: object
      required: [type, amount, currency, customerId]
      properties:
        type:
          type: string
          enum: [CREDIT_LIMIT]
        amount:
          type: number
          format: decimal
        currency:
          type: string
          example: "EUR"
        customerId:
          type: string

    ApprovalPolicyRule:
      type: object
      required: [type, policyName, approvers, requiredSignatures]
      properties:
        type:
          type: string
          enum: [APPROVAL_POLICY]
        policyName:
          type: string
        approvers:
          type: array
          items:
            type: string
        requiredSignatures:
          type: integer

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              timestamp: { type: string, format: date-time }
              status: { type: integer }
              error: { type: string }
              message: { type: string }
              path: { type: string }
    Conflict:
      description: Configuration conflict (e.g., rule already exists)
      content:
        application/json:
          schema:
            type: string
            example: "Cannot add rule: Rule already exists for key CUST-1"